Move RoutingTree into par module
The RoutingTree structure is only really useful when considering place and route
(after that only routing tables are really useful).
